financial_wellness/requirements.txt
flask
pandas
numpy
python-dotenv
pillow
pytesseract
streamlit
PyPDF2
cohere
fpdf
prophet
openpyxl
plotly
markdown
google-generativeai
requests
requests
flask
flask-cors
scikit-learn
joblib